apps/sync-service/pyproject.toml
[project]
name = "sync-service"
version = "1.0.0"
description = "FastAPI sync service for syncing images to Samsung Frame TV"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "websockets>=13",
    "aiohttp>=3.8.1",
    "async_timeout>=4.0.3",
    "Pillow>=10.0.0",
    "httpx>=0.24.0",
    "samsungtvws @ git+https://github.com/NickWaterton/samsung-tv-ws-api.git",
]

[project.optional-dependencies]
test = ["pytest>=7.4.0", "httpx>=0.24.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
